Made cleaning up on the chat provider
