+++ contexts/stream-wrapper.web.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const StreamWrapper: React.FC<any> = ({ children }) => {
+    return (
+        children
+    )
+}
+
+export default StreamWrapper
